+++ src/middlewared/middlewared/plugins/usage.py
@@ -272,8 +272,8 @@ class UsageService(Service):
         return {
             'platform': f'TrueNAS-{await self.middleware.call("system.product_type")}',
             'version': await self.middleware.call('system.version'),
-            'is_vendored': await self.middleware.call('system.vendor.is_vendored'),
-            'vendor_name': await self.middleware.call('system.vendor.name'),
+            'is_vendored': await self.call2(self.s.system.vendor.is_vendored),
+            'vendor_name': await self.call2(self.s.system.vendor.name),
             'is_virtualized': await self.call2(self.s.hardware.virtualization.is_virtualized),
             'hypervisor': await self.call2(self.s.hardware.virtualization.variant),
         }
